+++ src/Subscriber/CategoryView.php
@@ -20,7 +20,7 @@ class CategoryView implements EventSubscriberInterface
 
     private string $fieldName;
 
-    private array $initial;
+    private array $addParams;
 
     public function __construct(
         AbstractCategoryRoute $cmsPageRoute,
@@ -31,7 +31,7 @@ class CategoryView implements EventSubscriberInterface
         $this->cmsPageRoute = $cmsPageRoute;
         $this->categoryPath = $categoryPath;
         $this->fieldName    = $categoryPathFieldName;
-        $this->initial      = $initialNavigationParams;
+        $this->addParams    = $initialNavigationParams;
     }
 
     public static function getSubscribedEvents()
@@ -43,32 +43,19 @@ class CategoryView implements EventSubscriberInterface
     {
         $navigationId     = $event->getRequest()->get('navigationId', $event->getSalesChannelContext()->getSalesChannel()->getNavigationCategoryId());
         $category         = $this->cmsPageRoute->load($navigationId, $event->getRequest(), $event->getSalesChannelContext())->getCategory();
-        $path             = $this->getPath($category);
+
         $disableImmediate = safeGetByName($category->getCustomFields())(OmikronFactFinder::DISABLE_SEARCH_IMMEDIATE_CUSTOM_FIELD_NAME);
         $isHome           = $event->getRequest()->get('_route') === 'frontend.home.page';
-        $addParams        = $this->prepareAddParams();
-        $communication    = [
-            'search-immediate' => !$isHome && !$disableImmediate ? 'true' : 'false',
-            'category-page'    => $this->prepareCategoryPath($path),
-        ];
+        $isCategory       = !$isHome && !$disableImmediate;
 
-        if (!empty($addParams)) {
-            $communication['add-params'] = $addParams;
-        }
+        $categoryPath     = $this->getPath($this->cmsPageRoute->load($navigationId, $event->getRequest(), $event->getSalesChannelContext())->getCategory());
 
-        $event->getPage()->getExtension('factfinder')->assign(
-            [
-                'communication' => $communication,
-            ]);
-    }
+        $communication = [
+                'search-immediate' => !$isHome && !$disableImmediate ? 'true' : 'false',
+                'add-params'       => implode(',', $this->addParams),
+            ] + ($isCategory ? ['category-page' => $this->prepareCategoryPath($categoryPath)] : []);
 
-    private function prepareAddParams(): ?string
-    {
-        if (empty($this->initial)) {
-            return null;
-        }
-
-        return implode(',', $this->initial);
+        $event->getPage()->getExtension('factfinder')->assign(['communication' => $communication]);
     }
 
     private function prepareCategoryPath($path): string