+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-## Unreleased
+## [v2.2.2] - 2021.12.22
 ## Add
  - Category
    - Implement `ff-communication/category-page` attribute which serves to correctly filter products on category pages 
@@ -101,6 +101,7 @@
 ## [v1.0.0] - 2021.06.29
 Initial module release. Includes Web Components v4.0.3
 
+[v2.2.2]: https://github.com/FACT-Finder-Web-Components/shopware6-plugin/releases/tag/v2.2.2
 [v2.2.1]: https://github.com/FACT-Finder-Web-Components/shopware6-plugin/releases/tag/v2.2.1
 [v2.2.0]: https://github.com/FACT-Finder-Web-Components/shopware6-plugin/releases/tag/v2.2.0
 [v2.1.1]: https://github.com/FACT-Finder-Web-Components/shopware6-plugin/releases/tag/v2.1.1